+++ docs/docusaurus.config.js
@@ -5,6 +5,8 @@
 // See: https://docusaurus.io/docs/api/docusaurus-config
 import { themes as prismThemes } from 'prism-react-renderer'
 import smartyPants from 'remark-smartypants'
+import tailwindcss from 'tailwindcss'
+import autoprefixer from 'autoprefixer'
 
 /** @type {import('@docusaurus/types').Config} */
 const config = {
@@ -80,6 +82,20 @@ const config = {
     ],
   ],
 
+  plugins: [
+    async function tailwindPlugin() {
+      return {
+        name: 'docusaurus-tailwindcss',
+        configurePostCss(postcssOptions) {
+          // Appends TailwindCSS and AutoPrefixer.
+          postcssOptions.plugins.push(tailwindcss)
+          postcssOptions.plugins.push(autoprefixer)
+          return postcssOptions
+        },
+      }
+    },
+  ],
+
   markdown: {
     mermaid: true,
   },
@@ -126,6 +142,7 @@ const config = {
             label: 'Training',
           },
           { to: '/blog', label: 'Blog', position: 'left' },
+          { to: '/hub', label: 'MorioHub', position: 'right' },
           {
             href: 'https://github.com/certeu/morio',
             label: 'GitHub',